+++ src/Exceptions/FileCannotBeAdded/FileUnacceptableForCollection.php
@@ -2,7 +2,7 @@
 
 namespace Spatie\Medialibrary\Exceptions\FileCannotBeAdded;
 
-use Spatie\Medialibrary\Exceptions\FileCannotBeAdded;
+use Spatie\Medialibrary\Exceptions\FileCannotBeAdded\FileCannotBeAdded;
 use Spatie\Medialibrary\File;
 use Spatie\Medialibrary\HasMedia\HasMedia;
 use Spatie\Medialibrary\MediaCollection\MediaCollection;